requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-dotenv==1.0.0
SQLAlchemy==2.0.23
WTForms==3.1.1
redis==5.0.1
celery==5.3.4
requests==2.31.0
python-dateutil==2.8.2
click==8.1.7
python-socketio==5.10.0
eventlet==0.33.3